+++ devices/cuda/curtn.h
@@ -1,0 +1,16 @@
+// CURTN: a nano implementation of the CUDA Runtime API on top of the Driver API
+// Copyright 2024 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+#include <cuda.h>
+#include <cuda_runtime.h>
+
+namespace curtn
+{
+  // Unlike CUDA Runtime, CURTN requires explicit initialization before the first API call
+  cudaError_t init();
+
+  // Unlike CUDA Runtime, CURTN requires explicit initialization and cleanup of the current context
+  // Before releasing/destroying the context using the Driver API, this *must* be called
+  cudaError_t cleanupContext();
+}